backend/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.0.3
Flask==2.0.3
Flask-JWT-Extended==4.3.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
PyJWT==2.3.0
pyrsistent==0.18.1
python-decouple==3.6
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
zipp==3.7.0
